:sparkles: Criada rota para criacao de livro

+++ src/shared/infra/typeorm/migrations/1655334684717-Livros.ts
@@ -30,7 +30,7 @@ export class Livros1655334684717 implements MigrationInterface {
                     },
                     {
                         name: "price",
-                        type: "int"
+                        type: "float"
                     },
                     {
                         name: "created_at",
